chore: separate into `is-x` and `if-x` files
