+++ index.d.ts
@@ -76,9 +76,12 @@ export type {HasRequiredKeys} from './source/has-required-keys';
 export type {Spread} from './source/spread';
 export type {TupleToUnion} from './source/tuple-to-union';
 export type {IsEqual} from './source/is-equal';
-export type {IsAny, IfAny} from './source/is-if-any';
-export type {IsNever, IfNever} from './source/is-if-never';
-export type {IsUnknown, IfUnknown} from './source/is-if-unknown';
+export type {IsAny} from './source/is-any';
+export type {IfAny} from './source/if-any';
+export type {IsNever} from './source/is-never';
+export type {IfNever} from './source/if-never';
+export type {IsUnknown} from './source/is-unknown';
+export type {IfUnknown} from './source/if-unknown';
 
 // Template literal types
 export type {CamelCase} from './source/camel-case';
