+++ source/set-return-type.d.ts
@@ -1,4 +1,4 @@
-import type {IsUnknown} from './is-if-unknown';
+import type {IsUnknown} from './is-unknown';
 
 /**
 Create a function type with a return type of your choice and the same parameters as the given function type.
