+++ test-d/is-any.ts
@@ -1,5 +1,5 @@
 import {expectError, expectType} from 'tsd';
-import type {IsAny, IfAny} from '../index';
+import type {IsAny} from '../index';
 
 declare const anything: any;
 declare const something = 'something';
@@ -14,12 +14,5 @@ expectType<IsAny<unknown>>(false);
 expectType<IsAny<null>>(false);
 expectType<IsAny<undefined>>(false);
 
-// `IfAny` should return `true`/`false` if only `T` is specified
-expectType<IfAny<any>>(true);
-expectType<IfAny<any, 'T', 'F'>>('T');
-expectType<IfAny<string>>(false);
-expectType<IfAny<string, 'T', 'F'>>('F');
-
 // Missing generic parameter
 expectError<IsAny>(anything);
-expectError<IfAny>(anything);
